% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map.R
\name{add_polygon}
\alias{add_polygon}
\title{Add polygon to an invest map}
\usage{
add_polygon(map, investmap, country, highlight)
}
\arguments{
\item{map}{}

\item{investmap}{}

\item{country}{}

\item{highlight}{}
}
\value{
invest map
}
\description{
Add polygon to an invest map
}

