% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\name{adveqmap_options}
\alias{adveqmap_options}
\title{Set or get options for my adveqmap package}
\usage{
adveqmap_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for my adveqmap package
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{pal}}{(\code{list};2) The map palette }
 \item{\code{polygon_color}}{(\code{characater};"white") The value of polygon side color }
 \item{\code{url_tile}}{(\code{characater};"white") The value of url tile }
 \item{\code{style_list}}{(\code{characater};) The value of multibar plot font style }
 \item{\code{data_labels_theme}}{(\code{list};) multibar text settings}
 \item{\code{axis_labels_theme}}{(\code{list};) multibar axis settings}
}
}

