% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db.R, R/ratios.R
\name{get_fx}
\alias{get_fx}
\title{Get FX rate in USD and EUR}
\usage{
get_fx(db)

get_fx(db)
}
\arguments{
\item{db}{}

\item{db}{}
}
\value{
data.table of rates

data.table of rates
}
\description{
Get FX rate in USD and EUR

Get FX rate in USD and EUR
}

