% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ratios.R
\name{get_ratios}
\alias{get_ratios}
\title{Compute indicators used by map}
\usage{
get_ratios(db = load_data(...), level, ...)
}
\arguments{
\item{db}{data_base}

\item{level}{charcater level of aggregation. It can be
\itemize{
\item{country}
\item{region}
\item{subregion}
}}

\item{...}{other parametersused by load_data}
}
\value{
data.table containing
\itemize{
\item {region}
\item {subregion}
\item {multiple}
\item {fmv}
}
}
\description{
This functions computes indcators like fmv and multiple.
}

