% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db.R
\name{indicators}
\alias{indicators}
\title{Compute indicators used by map}
\usage{
indicators(db = load_data(...), ...)
}
\arguments{
\item{db}{data_base}

\item{...}{other parametersused by load_data}
}
\value{
data.table containing
\itemize{
\item {region}
\item {subregion}
\item {multiple}
\item {fmv}
}
}
\description{
This functions computes indcators like fmv and multiple.
}

