% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db.R
\name{load_data}
\alias{load_data}
\title{Load data from Excel}
\usage{
load_data(path, sheet = "RAW")
}
\arguments{
\item{path}{character Excel workbook path}

\item{sheet}{character sheet name default to RAW}
}
\value{
data.table containing
\itemize{
\item {location}{thegeography locatioin}
}
}
\description{
Load data from Excel
}

