% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/charts.R
\name{multiple_barplot}
\alias{multiple_barplot}
\title{Plot bar chart plot by underlying for a location}
\usage{
multiple_barplot(dat = load_data(), country, id_dom)
}
\arguments{
\item{dat}{data set}

\item{country}{character country name}
}
\value{
ndv3 plot discretBarChart
}
\description{
Plot bar chart plot by underlying for a location
}

