% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map.R
\name{plot.InvestMap}
\alias{plot.InvestMap}
\title{Plot investment map}
\usage{
\method{plot}{InvestMap}(x)
}
\arguments{
\item{x}{InvestMap object}
}
\description{
Plot investment map
}
\examples{
inMap <- InvestMap()
plot(inMap)
}

