% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map.R
\name{plot_map}
\alias{plot_map}
\title{Plot investment map}
\usage{
plot_map(ratios = get_ratios(), criteria = "multiple", currency = NULL)
}
\arguments{
\item{ratios}{data.table containing variables to plot}

\item{criteria}{character}

\item{currency}{character}
}
\value{
map
}
\description{
Plot investment map
}

