% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map.R
\name{popup_content}
\alias{popup_content}
\title{Create popup content}
\usage{
popup_content(country, nbr)
}
\arguments{
\item{country}{character country ou region selected}

\item{nbr}{numeric number of companies}
}
\value{
popup text as character
}
\description{
When a map shape is slected a popup can be trigged
}

