% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db.R
\name{load_data}
\alias{load_data}
\title{Load data from Excel}
\usage{
load_data(path, sheet = "RAW")
}
\arguments{
\item{path}{character Excel workbook path}

\item{sheet}{character sheet name default to RAW}
}
\value{
data.table containing
\itemize{
\item{loc_name} {:The location name}
\item{loc_id} {:The location id}
\item{region} {:The location region id}
\item{subregion} {:The location subregion id}
\item{fund} {:The fund name}
\item{under} {:The underyling name}
\item{part} {:The adveq participation}
\item{date} {:The date of investment}
\item{currency}{: The investment currency}
\item{industry} {:The industry sector}
\item{value} {:The investment amount}
\item{cost} {:The investment cost}
}
}
\description{
Load data from Excel
}

